+++ public/index.php
@@ -6,4 +6,17 @@
  * Time: 18:36
  */
 
-require_once __DIR__ . '/../app/runner.php';
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
+use Turbine\Application;
+
+require_once __DIR__ . '/../vendor/autoload.php';
+
+$app = new Application();
+
+$app->get('/', function (ServerRequestInterface $request, ResponseInterface $response) {
+    $response->getBody()->write('<h1>It works!</h1>');
+    return $response;
+});
+
+$app->run();